Changes request #1985
